ADNLPModelss

This package provides AD-based model implementations that conform to the NLPModels API. The following packages are supported:

ForwardDiff.jl: default choice.

Zygote.jl: you must load Zygote.jl separately and pass ADNLPModels.ZygoteAD as the adbackend keyword argument to the ADNLPModel or ADNLSModel constructor.

ReverseDiff.jl: you must load ReverseDiff.jl separately and pass ADNLPModels.ReverseDiffAD as the adbackend keyword argument to the ADNLPModel or ADNLSModel constructor.

Install

Install ADNLPModels.jl with the following command.

pkg> add ADNLPModels

Usage

This package defines two models, ADNLPModel for general nonlinear optimization, and ADNLSModel for nonlinear least-squares problems.

ADNLPModels.ADNLPModel — Type

ADNLPModel(f, x0)
ADNLPModel(f, x0, lvar, uvar)
ADNLPModel(f, x0, c, lcon, ucon)
ADNLPModel(f, x0, lvar, uvar, c, lcon, ucon)

ADNLPModel is an AbstractNLPModel using automatic differentiation to compute the derivatives. The problem is defined as

 min  f(x)
s.to  lcon ≤ c(x) ≤ ucon
      lvar ≤   x  ≤ uvar.

The following keyword arguments are available to all constructors:

minimize: A boolean indicating whether this is a minimization problem (default: true)

name: The name of the model (default: "Generic")

The following keyword arguments are available to the constructors for constrained problems:

lin: An array of indexes of the linear constraints (default: Int[])

y0: An inital estimate to the Lagrangian multipliers (default: zeros)

ADNLPModel uses ForwardDiff for the automatic differentiation by default. One can specify a new backend with the keyword arguments backend::ADNLPModels.ADBackend. There are three pre-coded backends:

the default ForwardDiffAD.

ReverseDiffAD accessible after loading ReverseDiff.jl in your environment.

ZygoteDiffAD accessible after loading Zygote.jl in your environment.

For an advanced usage, one can define its own backend and redefine the API as done in ADNLPModels.jl/src/forward.jl.

Examples

using ADNLPModels
f(x) = sum(x)
x0 = ones(3)
nvar = 3
ADNLPModel(f, x0) # uses the default ForwardDiffAD backend.

using ReverseDiff
ADNLPModel(f, x0; backend = ADNLPModels.ReverseDiffAD)

using Zygote
ADNLPModel(f, x0; backend = ADNLPModels.ZygoteAD)

using ADNLPModels
f(x) = sum(x)
x0 = ones(3)
c(x) = [1x[1] + x[2]; x[2]]
nvar, ncon = 3, 2
ADNLPModel(f, x0, c, zeros(ncon), zeros(ncon)) # uses the default ForwardDiffAD backend.

using ReverseDiff
ADNLPModel(f, x0, c, zeros(ncon), zeros(ncon); backend = ADNLPModels.ReverseDiffAD)

using Zygote
ADNLPModel(f, x0, c, zeros(ncon), zeros(ncon); backend = ADNLPModels.ZygoteAD)

ADNLPModels.ADNLSModel — Type

ADNLSModel(F, x0, nequ)
ADNLSModel(F, x0, nequ, lvar, uvar)
ADNLSModel(F, x0, nequ, c, lcon, ucon)
ADNLSModel(F, x0, nequ, lvar, uvar, c, lcon, ucon)

ADNLSModel is an Nonlinear Least Squares model using automatic differentiation to compute the derivatives. The problem is defined as

 min  ½‖F(x)‖²
s.to  lcon ≤ c(x) ≤ ucon
      lvar ≤   x  ≤ uvar

where nequ is the size of the vector F(x).

The following keyword arguments are available to all constructors:

linequ: An array of indexes of the linear equations (default: Int[])

minimize: A boolean indicating whether this is a minimization problem (default: true)

name: The name of the model (default: "Generic")

The following keyword arguments are available to the constructors for constrained problems:

lin: An array of indexes of the linear constraints (default: Int[])

y0: An inital estimate to the Lagrangian multipliers (default: zeros)

ADNLSModel uses ForwardDiff for the automatic differentiation by default. One can specify a new backend with the keyword arguments backend::ADNLPModels.ADBackend. There are three pre-coded backends:

the default ForwardDiffAD.

ReverseDiffAD accessible after loading ReverseDiff.jl in your environment.

ZygoteDiffAD accessible after loading Zygote.jl in your environment.

For an advanced usage, one can define its own backend and redefine the API as done in ADNLPModels.jl/src/forward.jl.

Examples

using ADNLPModels
F(x) = [x[2]; x[1]]
nequ = 2
x0 = ones(3)
nvar = 3
ADNLSModel(F, x0, nequ) # uses the default ForwardDiffAD backend.

using ReverseDiff
ADNLSModel(F, x0, nequ; backend = ADNLPModels.ReverseDiffAD)

using Zygote
ADNLSModel(F, x0, nequ; backend = ADNLPModels.ZygoteAD)

using ADNLPModels
F(x) = [x[2]; x[1]]
nequ = 2
x0 = ones(3)
c(x) = [1x[1] + x[2]; x[2]]
nvar, ncon = 3, 2
ADNLSModel(F, x0, nequ, c, zeros(ncon), zeros(ncon)) # uses the default ForwardDiffAD backend.

using ReverseDiff
ADNLSModel(F, x0, nequ, c, zeros(ncon), zeros(ncon); backend = ADNLPModels.ReverseDiffAD)

using Zygote
ADNLSModel(F, x0, nequ, c, zeros(ncon), zeros(ncon); backend = ADNLPModels.ZygoteAD)

Check the Tutorial for more details on the usage.

License

This content is released under the MPL2.0 License.

Contents

ADNLPModelss

Install

Usage

License

Contents

Reference

Contents

Index

Tutorial

ADNLPModel Tutorial

ADNLSModel tutorial