ADNLPModelss

This package provides a very simple model implement the NLPModels API. It uses ForwardDiff to compute the derivatives, which produces dense matrices, so it isn't very efficient for larger problems.

Install

Install ADNLPModels.jl with the following command.

pkg> add ADNLPModels

Usage

This package defines two models, ADNLPModel for general nonlinear optimization, and ADNLSModel other for nonlinear least-squares problems.

ADNLPModels.ADNLPModel — Type

ADNLPModel(f, x0)
ADNLPModel(f, x0, lvar, uvar)
ADNLPModel(f, x0, c, lcon, ucon)
ADNLPModel(f, x0, lvar, uvar, c, lcon, ucon)

ADNLPModel is an AbstractNLPModel using ForwardDiff to compute the derivatives. The problem is defined as

 min  f(x)
s.to  lcon ≤ c(x) ≤ ucon
      lvar ≤   x  ≤ uvar.

The following keyword arguments are available to all constructors:

name: The name of the model (default: "Generic")

The following keyword arguments are available to the constructors for constrained problems:

lin: An array of indexes of the linear constraints (default: Int[])

y0: An inital estimate to the Lagrangian multipliers (default: zeros)

ADNLPModels.ADNLSModel — Type

ADNLSModel(F, x0, nequ)
ADNLSModel(F, x0, nequ, lvar, uvar)
ADNLSModel(F, x0, nequ, c, lcon, ucon)
ADNLSModel(F, x0, nequ, lvar, uvar, c, lcon, ucon)

ADNLSModel is an Nonlinear Least Squares model using ForwardDiff to compute the derivatives. The problem is defined as

 min  ½‖F(x)‖²
s.to  lcon ≤ c(x) ≤ ucon
      lvar ≤   x  ≤ uvar

The following keyword arguments are available to all constructors:

linequ: An array of indexes of the linear equations (default: Int[])

name: The name of the model (default: "Generic")

The following keyword arguments are available to the constructors for constrained problems:

lin: An array of indexes of the linear constraints (default: Int[])

y0: An inital estimate to the Lagrangian multipliers (default: zeros)

Check the Tutorial for more details on the usage.

License

This content is released under the MPL2.0 License.

Contents

ADNLPModelss

Install

Usage

License

Contents

Reference

Tutorial

ADNLPModel Tutorial

ADNLSModel tutorial