Reference

​

Contents

​

Reference

Contents

Index

​

Index

​

ADNLPModels.ADNLPModel

ADNLPModels.ADNLSModel

​

ADNLPModels.ADNLPModel — Method

ADNLPModel(f, x0)
ADNLPModel(f, x0, lvar, uvar)
ADNLPModel(f, x0, c, lcon, ucon)
ADNLPModel(f, x0, lvar, uvar, c, lcon, ucon)

ADNLPModel is an AbstractNLPModel using automatic differentiation to compute the derivatives. The problem is defined as

 min  f(x)
s.to  lcon ≤ c(x) ≤ ucon
      lvar ≤   x  ≤ uvar.

The following keyword arguments are available to all constructors:

name: The name of the model (default: "Generic")

The following keyword arguments are available to the constructors for constrained problems:

lin: An array of indexes of the linear constraints (default: Int[])

y0: An inital estimate to the Lagrangian multipliers (default: zeros)

ADNLPModel uses ForwardDiff for the automatic differentiation by default. One can specify a new backend with the keyword arguments adbackend::ADNLPModels.ADBackend. There are three pre-coded backends:

the default ForwardDiffAD.

ReverseDiffAD accessible after loading ReverseDiff.jl in your environment.

ZygoteDiffAD accessible after loading Zygote.jl in your environment.

For an advanced usage, one can define its own backend and redefine the API as done in ADNLPModels.jl/src/ad.jl.

Examples

using ADNLPModels
f(x) = sum(x)
x0 = ones(3)
nvar = 3
ADNLPModel(f, x0) # uses the default ForwardDiffAD backend.

using ReverseDiff
ADNLPModel(f, x0; adbackend = ADNLPModels.ReverseDiffAD(nvar, f, x0))

using Zygote
ADNLPModel(f, x0; adbackend = ADNLPModels.ZygoteAD(nvar, f, x0))

using ADNLPModels
f(x) = sum(x)
x0 = ones(3)
c(x) = [1x[1] + x[2]; x[2]]
nvar, ncon = 3, 2
ADNLPModel(f, x0, c, zeros(ncon), zeros(ncon)) # uses the default ForwardDiffAD backend.

using ReverseDiff
reversediff_backend = ADNLPModels.ReverseDiffAD(nvar, ncon, f, x0)
ADNLPModel(f, x0, c, zeros(ncon), zeros(ncon); adbackend = reversediff_backend)

using Zygote
zygote_backend = ADNLPModels.ZygoteAD(nvar, ncon, f, x0)
ADNLPModel(f, x0, c, zeros(ncon), zeros(ncon); adbackend = zygote_backend)

ADNLPModels.ADNLSModel — Method

ADNLSModel(F, x0, nequ)
ADNLSModel(F, x0, nequ, lvar, uvar)
ADNLSModel(F, x0, nequ, c, lcon, ucon)
ADNLSModel(F, x0, nequ, lvar, uvar, c, lcon, ucon)

ADNLSModel is an Nonlinear Least Squares model using automatic differentiation to compute the derivatives. The problem is defined as

 min  ½‖F(x)‖²
s.to  lcon ≤ c(x) ≤ ucon
      lvar ≤   x  ≤ uvar

where nequ is the size of the vector F(x).

The following keyword arguments are available to all constructors:

linequ: An array of indexes of the linear equations (default: Int[])

name: The name of the model (default: "Generic")

The following keyword arguments are available to the constructors for constrained problems:

lin: An array of indexes of the linear constraints (default: Int[])

y0: An inital estimate to the Lagrangian multipliers (default: zeros)

ADNLSModel uses ForwardDiff for the automatic differentiation by default. One can specify a new backend with the keyword arguments adbackend::ADNLPModels.ADBackend. There are three pre-coded backends:

the default ForwardDiffAD.

ReverseDiffAD accessible after loading ReverseDiff.jl in your environment.

ZygoteDiffAD accessible after loading Zygote.jl in your environment.

For an advanced usage, one can define its own backend and redefine the API as done in ADNLPModels.jl/src/ad.jl.

Examples

using ADNLPModels
F(x) = [x[2]; x[1]]
nequ = 2
x0 = ones(3)
nvar = 3
ADNLSModel(F, x0, nequ) # uses the default ForwardDiffAD backend.

using ReverseDiff
ADNLSModel(F, x0, nequ; adbackend = ADNLPModels.ReverseDiffAD(nvar, x -> sum(F(x) .^ 2), x0))

using Zygote
ADNLSModel(F, x0, nequ; adbackend = ADNLPModels.ZygoteAD(nvar, x -> sum(F(x) .^ 2), x0))

using ADNLPModels
F(x) = [x[2]; x[1]]
nequ = 2
x0 = ones(3)
c(x) = [1x[1] + x[2]; x[2]]
nvar, ncon = 3, 2
ADNLSModel(F, x0, nequ, c, zeros(ncon), zeros(ncon)) # uses the default ForwardDiffAD backend.

using ReverseDiff
reversediff_backend = ADNLPModels.ReverseDiffAD(nvar, ncon, x -> sum(F(x) .^ 2), x0)
ADNLSModel(F, x0, nequ, c, zeros(ncon), zeros(ncon); adbackend = reversediff_backend)

using Zygote
zygote_backend = ADNLPModels.ZygoteAD(nvar, ncon, x -> sum(F(x) .^ 2), x0)
ADNLSModel(F, x0, nequ, c, zeros(ncon), zeros(ncon); adbackend = zygote_backend)